requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.0
python-dotenv>=1.0.0
requests>=2.31.0
gigachain>=0.1.0
cryptography>=41.0.0
langchain>=0.1.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
PyPDF2>=3.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pymupdf>=1.23.0
tqdm>=4.65.0